RISC-V SMP: fix initialization of PLIC threshold for APs
The current code is initializing the PLIC threshold for all APs
from the start_secondary_cores() function, executed by the boot
processor. However, PLIC configuration for a given CPU can be
changed by the boot ROM executed by the CPU at startup, and it has
been observed that the OpenSBI firmware changes the PLIC threshold
value, thereby overwriting the configuration set by the boot
processor.
This change fixes the above issue by moving the configuration of
the PLIC threshold for a given CPU to the ap_start() function,
executed by the CPU itself.