RISC-V SMP: fix parsing of CPU device tree entries
In the RISC-V SMP initialization code, the device tree is parsed in
order to retrieve the number of present CPUs. Commit
364b517d03a7296e8a5cb0326292d87efac44e07 introduced a regression by
which a cpu device tree entry, whose name is null-terminated in the
device tree, can no longer be recognized, thereby preventing the
kernel from enumerating the application processors.
This change fixes the above regression by using the
sstring_from_cstring() function when parsing a device tree entry
name.