LDFLAGS: add `-z notext` and `-z noexecstack` linker flags
Starting with release 2.39 of the GNU Binutils, the ld linker warns
with "missing .note.GNU-stack section implies executable stack" if
it encounters an object file without a .note.GNU-stack section, and
with "relocation in read-only section" if it encounters an object
file containing a text relocation. In the both the kernel and the
various bootloaders, such object files exist and are those built
from assembly source files.
The `-z noexecstack` linker option silences the first warning (and
makes the linker assume that the output executable does not need an
executable stack).
The `-z notext` option silences the second warning (which does not
reveal an actual issue, because the kernel is able to apply text
relocations when doing KASLR).

https://github.com/nanovms/nanos/issues/1870