x86 bootloader: silence linker warnings and fix output sections
Starting with release 2.39 of the GNU Binutils, the linker warns
when creating an ELF file containing a LOAD segment with RWX
permissions (https://github.com/nanovms/nanos/issues/1870). With
the curent linker script for the x86 bootloader, the resulting ELF
file contains such a segment.
This change adds a PHDRS directive to the linker script and maps
the different sections to the appropriate segments, so that no
LOAD segments with RWX permissions exist in the stage2.elf file.
This does not affet the final binary file of the bootloader, but
silences the above warning.
In addition, the SECTIONS directive of the linker script has been
fixed so that all the input sections are mapped to the output
sections specified in the script; this decreases by around 2% the
size of the bootloader binary file.

Closes #2010