Klib linker scripts: use ELF segments with appropriate permissions
This change amends the linker scripts for the klibs so that each
klib ELF file has 3 segments: a text segment with read and execute
permissions, a data segment with read-only permissions, and a data
segment with read/write permissions. This avoids mapping an entire
klib image with both write and execute permissions when the kernel
loads a klib.
In addition, the SECTIONS directive in the linker scripts has been
fixed so that all the read-only input sections are mapped to the
.text and .rodata output sections; this decreases the size of the
klib ELF files.