tls input buffer handler: fix check for application layer error
When a TLS handshake with a remote peer is complete, the TLS input
buffer handler invokes the application layer connection handler,
which returns the application layer input buffer handler for the
connection. Any error at this stage should be reported by the
application layer by returning INVALID_ADDRESS; this is consistent
with the behavior for non-encryped connections (see
direct_receive_service() in net/direct.c), and allows applications
to not implement an input buffer handler (e.g. when they connect to
a remote peer to only send data and then close the connection), in
which case their connection handler can return 0.
This change modifies the TLS input buffer handler so that the check
for errors uses INVALID_ADDRESS instead of 0, and modifes the gcp
and cloudwatch code to align with this implementation.