http_request(): insert Content-length header in all requests
This change makes http_request() insert the Content-length HTTP
header in any request, regardless of the presence of a non-empty
request body. This is necessary in order to support some types of
requests which require this header (for example PUT requests to
the Azure blob storage service to create a blob).