Klibs: fix handling of missing dependencies
The kernel code that automatically loads klibs found in the /klib
folder has a flaw by which a klib that at a first attempt fails to
initialize due to missing dependencies is put in a state where it
cannot be initialized even after these dependencies are satisfied
by other klibs that are subsequently loaded. This is because the
`pending` variable cannot be safely used in a lock-free manner to
determine whether any other klibs are about to be initialized; for
example, in an SMP VM one core could set `pending` to 0 and another
core could put a klib in a failed state before the first core
initializes the just loaded klib.
This issue is causing sporadic CI test failures, such as
https://app.circleci.com/pipelines/github/nanovms/nanos/4623/workflows/f30b3b7f-0732-49d6-9e09-f9efbd5d6e21/jobs/16230.

This change fixes the above issue by introducing a
spinlock-protected klib_autoload structure that keeps track of
pending klibs and loaded klibs with missing dependencies. As a side
effect, the lock in this structure protects the `klib_loaded`
vector and the global kernel symbol table from concurrent
modifications.