buffer_set_capacity(): add handling of buffers without contents
This change makes the buffer_set_capacity() function work with
buffers without contents (i.e. buffer structs with a zero `length`
field). This allows buffers initialized via `init_buffer()` to be
used as dynamically allocated buffers without having to do an
initial allocation for buffer contents.
The `buffer_set_capacity()` function is being moved from a header
file to a source file because it can be computationally intensive
(due to memory allocation and deallocation operations, as well as
memory copying) and as such should not be called in hot code paths.
This decreases the kernel binary size by about 55 KB.