Stage3: initialize Unix subsystem before loading klibs
A given klib may make modifications to the implementation of Unix
syscalls, and for this it needs the Unix subsystem to be
initialized (an example of this is the sandbox klib).
This change moves the initialization of the Unix subsystem from
startup() to create_init() in stage3.c (which is called immediately
after the root filesystem is loaded). This fixes an "assertion
m[n].handler == 0 failed at /usr/src/nanos/src/unix/syscall.c:2828
(IP 0xffffffffa9ca5100) in _register_syscall()" which can happen in
SMP instances if the sandbox klib is loaded before the startup()
function in stage3.c is called.