sched_setaffinity(): add check for valid CPU set
If the CPU set supplied to sched_setaffinity() does not include any
online CPU in the machine, this syscall should return -EINVAL.
In addition, the copy from user memory is being fixed so that it
does not overrun the user buffer when cpusetsize is not a multiple
of 8 bytes.
The affinity bitmap is being moved from struct thread to struct
sched_task, in preparation for changing the scheduler so that it
honors CPU affinity.