scheduler: add support for CPU affinity (thread pinning)
This change enhances the scheduler so that it honors the CPU
affinity of user threads (set via the sched_setaffinity syscall).

The migrate_to_self() function is being amended so that it does not
fail to wake up an idle CPU if the current CPU could not migrate a
thread from the idle CPU (because threads may be non-migratable due
to their CPU affinity).
Since a task dequeued from a scheduling queue may not be the
highest priority (lowest runtime) task in that queue, its runtime
field may be non-zero; the thread_cputime_update() function is
being updated to reflect that.