Unix sockets: fix lock inversion with notify sets
Notify set interface functions such as notify_dispatch() and
notify_remove() (which internally acquire the notify set lock)
should never be called while holding the same lock that is acquired
in the notify handler (which is invoked with the notify set lock
held), because this can lead to a lock inversion and ensuing
deadlock if the notify handler is invoked on a given vCPU at the
same time as another vCPU calls a notify set function.
This change fixes the Unix domain socket connect code to avoid the
above issue.
Note: in order to prevent notify set functions from being called
with a notify entry that has been deallocated (by another vCPU),
notify entries should be refcounted.