RISC-V SMP: fix update of number of online vCPUs
In the RISC-V architecture, secondary cores are started in parallel
(i.e. the next core is started without waiting for the previous
core startup process to complete), and the boot core enters the
runloop without waiting for secondary cores to start. This means
that the scheduler can run before all cores are online. Since
secondary cores can start in an arbitrary order, the cpuinfos array
is also populated in an arbitrary order; The scheduler assumes that
all vCPUs with id from 0 to (`total_processors` - 1) are online,
therefore if each vCPU increments `total_processors` by 1 when
starting up it may happen that the scheduler tries to access a null
cpuinfo. This issue is fixed by amending `ap_start_newstack()` so
that it updates `total_processors` based on which vCPUs have
completed the startup process, so that the above assumption in the
scheduler is always valid.
In addition, if the cpuinfos vector is initially allocated with
length=1 and then extended as secondory cores are brought online,
the scheduler may access invalid vector elements, because the
backing buffer of the vector may change at any time due to a vector
extension. This issue is fixed by amending `init_kernel_contexts()`
so that it allocates the vector with a capacity set to
`present_processors`: this ensures that the backing buffer never
changes.
In `kernel_runtime_init()`, `count_cpus_present()` is now called
before `init_kernel_contexts()`, so that the number of present
vCPUs is known when allocating the cpuinfos vector.
In the x86 architecture, the ACPI tables are now initialized in
`count_cpus_present()` instead of `init_interrupts()`, because they
are needed in order to be able to enumerate the vCPUs.