file_open(): refactor using new function unix_file_new()
The new function unix_file_new() allocates a file descriptor for an
open file; it can be used when some operations have to be performed
on an open file before creating a file descriptor (and thus making
the file available to userspace).
This function will be used by memfd_create() when adding support
for file sealing.