memfd_create(): add support for file sealing
This change enhances the tmpfs implementation so that it supports
file sealing, and modifies memfd_create() so that it acts upon the
MFD_ALLOW_SEALING flag when configuring the initial set of seals
for the underlying tmpfs file.
The fcntl() syscall now accepts 2 additional commands, needed for
file sealing operations: F_ADD_SEALS and F_GET_SEALS.
The fcntl runtime test has been amended to exercise this new
feature.