poll/select: fix race condition
When poll() is invoked by the user program, the kernel allocates a
new epoll_blocked struct and associates it to the thread-specific
epoll instance for poll/select; then, the set of supplied file
descriptors is processed, and for each file descriptor the relevant
event conditions are evaluated; finally, the thread is blocked if
no events of interest are pending. This logic presents a race
condition when one or more of the supplied file descriptors is
already registered in the epoll instance: from the moment the
epoll_blocked struct is associated to the epoll instance, the
notify entry handler (which can be invoked at any time for all
currently registered file descriptors) can modify the user buffers
referenced by the epoll_blocked struct, which may still be in the
process of being initialized by the thread calling poll(); this can
lead to inconsistencies between the contents of these buffers and
the return value stored in the epoll_blocked struct.
A possible symptom of the above issue is the assertion failure
reported in #1963; another possible symptom is a polling thread
that remains stuck indefinitely even though the events being polled
for have been notified (this has been observed when running the
inotify runtime test on a 2-vCPU instance).

A similar race condition affect the select() implementation.

Ths change fixes the above issues by moving the allocation of the
epoll_blocked struct after the user buffers passed to the poll() or
select() syscall have been processed; the check for pre-existing
event conditions is now done right after allocating the
epoll_blocked struct, via a newly defined function
epoll_check_epollfds() which consolidates code shared between
poll(), select() and epoll_wait().

In addition, the lock that protects the epollfd list in an epoll
instance in not taken anymore when poll() or select() is called,
because these syscalls operate on a thread-local epoll instance,
whose epollfd list is never touched concurrently by other threads.