Robust futexes: fix handling of robust list
When a thread terminates while holding a mutex, the kernel signals
to any mutex waiters that the mutex holder has terminated, by
processing the robust list of the terminated thread and executing a
FUTEX_WAKE operation for each futex in the list. When the mutex
waiter thread resumes, the NPTL code in glibc frees the relevant
userspace mutex data structure, which means that the next pointer
in the robust list may not be valid anymore; if the CPU that is
handling the thread termination accesses the next pointer after
another CPU has started executing the woken up thread, it may
retrieve a bogus pointer value, which can lead to a crash in
userspace; this issue has been observed when running the
futexrobust runtime test on RISC-V instances with multiple vCPUs.

This change fixes the above issue by retrieving the pointer to the
next entry in the robust list before waking up any waiter on the
current entry.
In addition, handling of the `list_op_pending` pointer is being
fixed so that the mutex in the pending state, if present, is not
processed while walking the robust list; this fix prevents the
mutex from being processed twice.