On-demand program file paging: fix initialization of BSS areas
When on-demand paging of the program file is enabled, BSS areas in
pages faulted-in on demand are zeroed in-place, i.e. a newly mapped
page retrieved via the page cache is zeroed starting from the BSS
offset set up when initializing the relevant vmap. This creates a
problem if the page contains other data (e.g. from another loadable
section of the program) at or after the BSS offset, in which case
this data would be overwritten.
This change fixes the above issue by using a separate page (instead
of the page from the page cache) where the initialized program data
(located before the BSS offset) is copied from the page cache page,
and the rest of the page (starting at the BSS offset) is zeroed
out. Closes https://github.com/nanovms/ops/issues/1629.

The on-demand paging implementation is being reworked to address
the following shortcomings:
- A vmap struct cannot be referenced without holding the vmap lock,
because it may be modified and/or deallocated at any time (e.g.
if the access protection flags of contiguous memory areas are
modified)
- Parallel handling of page faults from different CPUS for the
same page cannot be safely handled via a process-global pending
fault list, because it is possible for a faulting CPU to create a
new pending fault for a given page and then complete it before
another faulting CPU processes a fault for the same page: in this
case, the second CPU would not find any pending fault for the page,
even though the fault has been handled by the first CPU
The reworked implementation allows multiple faults to be pending
simultaneously for the same page, and relies on the page table lock
to prevent multiple mappings of the same page by different CPUs.