Page fault: fix handling of faults in kernel contexts
It is possible for a page fault to occur in a kernel context: for
example, when setting up a signal frame on the stack of a user
thread, if the relevant page of the stack is not resident in memory
it will cause a fault. In these cases, if the fault cannot be
resolved synchronously (e.g. because a filesystem sync to disk is
needed to free up some memory), the existing code panics at an
assert(!is_kernel_context(ctx)) in the interrupt handler triggered
by the fault.
This change removes the above assert() so that page faults in kernel
contexts are correctly processed.