vmap_update_protections_intersection(): make more generic
This function, which is currently being used only to update the
protection flags of a memory range, is being renamed to
vmap_update_flags_intersection() and made more generic so that it
can be used to update an arbitrary set of flags.
In the next commit this function will be used by madvise().