Memory mapping: add support for Transparent Huge Pages (THP)
Memory areas mapped in the user process virtual address space are
paged on-demand when the process causes a page fault by
dereferencing a virtual address.
In the existing implementation, a single 4-KB memory page is
allocated and mapped after a page fault; with THP, memory pages
belonging to anonymous mappings can be allocated and mapped using
larger sizes (up to 2 MB); this decreases the number of page faults
triggered when the process accesses a large memory area, resulting
in improved performance.
The THP feature is enabled by default in memory areas mapped via
the mmap() syscall, and can be enabled and disabled dynamically for
an arbitrary (non-file-backed) memory area via the newly
implemented madvise() syscall, by using the MADV_HUGEPAGE and
MADV_NOHUGEPAGE advice values.

In order to be able to allocate more than 4 KB at a time, the vmap
heap is being changed from the `pages` kernel heap to a new heap
that allocates directly from the `page_backed` heap. The issue that
prompted the use of the `pages` heap (i.e. excessive memory
fragmentation when the process makes heavy use of mmap and munmap)
is no longer an issue because allocating larger (and always
power-of-2) chunks of memory at a time decreases fragmentation.