exec_elf_finish(): add setting of cwd for non-static executables
Configuration of the current working directory via the `cwd`
attribute of the manifest was being done only for statically linked
program executables. This change enables the `cwd` attribute to
take effect also for non-static executables.
In addition, a missing deallocation of the executable file buffer
for non-static executables is being fixed.

Closes https://github.com/nanovms/ops/issues/1632