Page tables: remove completions from mapping functions
Asynchronous completions were previously needed to properly handle
page faults in SMP instances, so that a pending fault could be
released only after all vCPUS executed a TLB shootdown. Since TLB
shootdowns are only needed when an existing mapping is either
modified or deleted, and are not needed when a new mapping is
created, the current kernel no longer waits for TLB shootdown
completion before releasing a pending fault, and instead relies on
proper locking to ensure that the same virtual address is not
mapped concurrently by different vCPUs. This new approach removes
the need for asynchronous completions in the mapping functions; in
addition, completions are not needed in the TLB shootdown code.
This change removes completions for mapping and TLB shootdown
functions, and modifies the map() function so that any existing
mapping is always overwritten (possibly splitting existing
block-level mappings into page-level mappings).
The flags_from_pte() function is being amended so that the returned
flags do not include the flag that differentiates between block and
page mappings; this allows flags retrieved from a block mapping to
be used by the pte_split() function when creating a page mapping.