Id heap dealloc_from_range(): ensure alignment of next_bit values
When support for block-level mappings in userspace memory is added,
it may happen that a physical memory range being deallocated is not
aligned to a power-of-2 value (e.g. when a 2MB mapping is partially
unmapped); in this scenario, the id heap may set the next_bit field
corresponding to the deallocation size to a non-aligned value,
which may cause subsequent allocations of the same size to return a
non-aligned value, thus breaking the property of the id heap that
allocated values are aligned to the nearest power-of-2 value equal
to or greater than the allocation size.

This change fixes the above issue in the id heap by adding a check
on the bit range being deallocated before setting the next_bit
field.