ARM: add support for different physical memory layouts
With this change, it is now possible to run Nanos on platforms
(such as Azure ARM-based Standard_B2pts_v2 instances) where the
range of memory addresses in the kernel ELF file
(0x40400000 - 0x40600000) is not a valid physical memory range; in
these cases, the kernel is loaded at physical addresses different
from the ELF addresses. The offset between the kernel location in
physical memory and the ELF addresses is stored in the
`kernel_phys_offset` global variable.
In addition, the address range for memory-mapped I/O is now
dynamically detected when running on UEFI platforms, instead of
being hard-coded to (0x00000000 - 0x40000000).