pci_parse_iomem(): fix detection of BARs with 64-bit address
The current code tries to detect if a BAR is implemented by
checking if the address register is zero after masking the flag
bits; however, if a 64-bit memory BAR address has the lower 32 bits
set to zero, this approach wronlgy considers the bar as
non-implemented, and fails to parse the next BAR as the upper 32
bits. The correct way to check if a BAR is implemented is by
comparing the address register value with zero, without masking the
flag bits.
This change fixes a "set_gap error" meassge that is printed on the
console when running an ARM instance using a QEMU "virt" machine
with the "highmem" setting enabled and with UEFI firmware.