ARM: add parsing of device tree blob
This change amends the init_platform_devices() function to add
parsing of the device tree in order to retrieve the number (and
identifiers) of the vCPUS and the base address of the PCI
configuration space. This allows the kernel to properly detect the
number of vCPUs when ACPI tables are not present (e.g. when running
an on-prem instance with Ops) and to run without the 'highmem=off'
Qemu command line option.
In order to be able to parse the device tree during initialization,
its backing memory is marked as reserved when initializing the
physical memory heap, and is reclaimed only after parsing the
device tree.
The device tree parser has been enhanced with a number of functions
that allow retrieving multiple pieces of information from a device
tree without having to re-parse the blob from the beginning
multiple times. To this end, the fdt struct has been defined: this
struct, which is initialized by dtb_parse_init(), holds information
that is needed during the parsing, and keeps track of the current
position of the parser in the device tree blob.