Scheduler timer period: remove capping to RUNLOOP_TIMER_MAX_PERIOD
Setting the platform timer to wake up the kernel when there are no
kernel timers to be dispatched causes useless CPU usage, which is
especially noticeable when the system is idle.
This change removes the capping of the scheduler timer period to
RUNLOOP_TIMER_MAX_PERIOD_US, so that unneeded timer interrupts are
avoided; the RUNLOOP_TIMER_MAX_PERIOD_US setting is still used
when scheduling user threads, which prevents a thread from running
for too long and starving out other threads. In addition, the min
and max fields of struct timerqueue are being removed because their
usage is specific to the scheduler timer queue.