LWIP timers: remove custom implementation in Nanos
Using a statically configured set of timer intervals causes useless
CPU usage when the timers do not need to be dispatched, especially
with high-frequency timers.
This change removes the LWIP timer definitions from Nanos, and uses
the timer management code inside LWIP; this allows dynamic
enablement of some timers (specifically, the TCP and MLD6 timers),
which decreases CPU usage (especially noticeable when the system is
idle).