Filesystem: add DMA handling to file I/O closures
Calls to DMA functions are being moved from the pagecache code to
the generic filesystem code, in preparation for a future commit
which will allow using filesystems without a page cache.
In order to decrease the number of closure allocations, 2 new
callback functions (file_read and file_write) have been added to
struct filesystem: the implementation of these functions expects to
be passed DMA buffers, while the read and write closures stored in
struct fsfile operate on normal memory buffers.