Filesystem: allow use without page cache
It is now possible to use a filesystem without the page cache.
All bootloaders (legacy PC and UEFI) and tools (mkfs, dump,
tfs-fuse) ae now built without the pagecache code; this allows
avoiding unnecessary memory copying when reading from or writing to
the filesystem, and decreases binary file sizes, thereby speeding
up the boot process.
`#ifdef KERNEL` pre-processor directives have been moved from the
pagecache code (which is now exclusively for kernel use) to the
filesystem code (which can now by used by code that does not use a
page cache).
When reading from a filesystem, populating the scatter-gather list
is now done by the calling code (which does not do any copying of
memory buffers), and the SG list is consumed inside the read
implementation (i.e., either the pagecache code, or the driver of
the underlying storage device).
In order to support the sendfile() syscall (which does not supply
any memory buffers when reading from the input file), the
pagecache_node_fetch_pages() function has been amended to take
optional SG list and completion arguments, so that the calling code
can access a populated SG list when the completion is invoked.
filesystem_read_linear() and filesystem_write_linear() have been
enhanced to convert unaligned buffers to aligned buffers, which is
required in order to interface directly with filesystem code
without going through the page cache.
sg_file_io closures in the fdesc structure have been replaced by
file_iov closures, which take a struct iovec array instead of an
sg_list; this allows eliminating conversions between struct iovec
arrays and SG lists when SG lists are not needed (as in the network
socket and Unix domain socket implementations), and will allow
creating file-descriptor-specific SG lists (which may not
necessarily be a 1-to-1 mapping between struct iovec entries and
SG list buffers) when SG lists are needed (as in the regular file
I/O implementation).