ARM PCI: properly handle missed detection of ECAM base address
This change amends the pci_cfgread() and pci_cfgwrite() functions
so that they don't do any MMIO access if the base address of the
PCI configuration space has not been detected. This prevents access
to invalid memory addresses in platforms (such as Azure ARM-based
instances) where the ECAM base address is not exposed via the
device tree or the MCFG ACPI table.