kernel_shutdown(): sync storage after shutdown completions are run
Some shutdown completions (such as those implemented in the syslog
klib, in the coredump code, and in the tracelog code) write data to
disk; therefore, the storage_shutdown closure should be invoked
only after all shutdown completions have finished their job,
otherwise data written by a shutdown completion may not be flushed
to disk. In addition, any error status encountered when executing a
shutdown handler or writing data to disk should be printed to the
console.