Core dump: fix double refcount release of core dump file
The current code is calling fsfile_release() twice for the file
holding a core dump; this causes the fsfile struct to be
deallocated before file data is flushed to disk, and as a result
the core dump file in the disk contains all zeros.
This change fixes the above issue by removing the unnecessary
dump_complete() closure, which was calling fsfile_release().
The core dump test has been made more robust by adding a file size
check, so the test will fail if a zero-sized core dump file is
generated.