9P filesystem: fix operation with deleted and renamed files
A given file can be deleted and renamed while still open, and all
I/O operations should still be possible after file deletion or
renaming.
This change fixes the 9P filesystem driver so that it meets the
above requirements:
- the SETATTR command to set the file size before writing to a file
has been removed, because it doesn't work (the server returns an
error) if the file has been deleted or renamed
- deletion of the dentry structure associated to a file has been
removed from the functions that delete or rename the file, so that
an open file can still be used after deletion or renaming
- on file closure, a check on the existence of the file metadata
tuple has been added, so that access to non-existing metadata is
avoided

Fixes #2047.